requirements.txt
tensorflow
librosa
numpy
pandas
pydub
python-dotenv